# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AAA Bets - Horse Racing Predictor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: races
    description: Race operations
  - name: predictions
    description: Prediction operations
  - name: weights
    description: Weight configuration operations
  - name: chat
    description: AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /races:
    get:
      operationId: getRaces
      tags: [races]
      summary: Get races with weekly forecast context
      parameters:
        - name: venue
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of races
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Race"
    post:
      operationId: createRace
      tags: [races]
      summary: Create a race
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRaceBody"
      responses:
        "201":
          description: Race created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"

  /races/{raceId}:
    get:
      operationId: getRace
      tags: [races]
      summary: Get race by ID
      parameters:
        - name: raceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaceDetail"
        "404":
          description: Race not found

  /races/{raceId}/horses:
    get:
      operationId: getRaceHorses
      tags: [races]
      summary: Get horses in a race
      parameters:
        - name: raceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of horses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Horse"
    post:
      operationId: addHorse
      tags: [races]
      summary: Add a horse to a race
      parameters:
        - name: raceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHorseBody"
      responses:
        "201":
          description: Horse added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Horse"

  /races/{raceId}/predictions:
    get:
      operationId: getRacePredictions
      tags: [predictions]
      summary: Get persisted predictions for a race
      parameters:
        - name: raceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Predictions ranked by score
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"

  /races/{raceId}/analyze:
    post:
      operationId: analyzeRace
      tags: [predictions]
      summary: Trigger forecast generation for a race
      parameters:
        - name: raceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Forecast generated and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Race not found

  /races/{raceId}/result:
    post:
      operationId: recordRaceResult
      tags: [predictions]
      summary: Record an official result and grade the forecast
      parameters:
        - name: raceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordRaceResultBody"
      responses:
        "201":
          description: Result recorded and grading completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaceResultSummary"

  /weights:
    get:
      operationId: getWeights
      tags: [weights]
      summary: Get current prediction weights
      responses:
        "200":
          description: Current weights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionWeights"
    put:
      operationId: updateWeights
      tags: [weights]
      summary: Update prediction weights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWeightsBody"
      responses:
        "200":
          description: Updated weights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionWeights"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message to the analyst
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: AI response with possible weight adjustments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get recent chat history
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [races]
      summary: Get dashboard summary stats and weekly outlook
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    PredictionFactors:
      type: object
      properties:
        courseForm:
          type: number
        formDistance:
          type: number
        jockeyTrainer:
          type: number
        oddsMovement:
          type: number
        history:
          type: number
        overall:
          type: number
      required:
        - courseForm
        - formDistance
        - jockeyTrainer
        - oddsMovement
        - history
        - overall

    PredictionSummary:
      type: object
      properties:
        id:
          type: integer
        horseId:
          type: integer
        horseName:
          type: string
        rank:
          type: integer
        score:
          type: number
        baseConfidence:
          type: number
        confidence:
          type: number
        confidenceDelta:
          type: number
        confidenceBand:
          type: string
        timeToRaceMinutes:
          type: ["integer", "null"]
        resultStatus:
          type: string
        finishPosition:
          type: ["integer", "null"]
        aiSummary:
          type: ["string", "null"]
      required:
        - id
        - horseId
        - horseName
        - rank
        - score
        - baseConfidence
        - confidence
        - confidenceDelta
        - confidenceBand
        - resultStatus

    Prediction:
      allOf:
        - $ref: "#/components/schemas/PredictionSummary"
        - type: object
          properties:
            raceId:
              type: integer
            factors:
              $ref: "#/components/schemas/PredictionFactors"
            createdAt:
              type: string
            gradedAt:
              type: ["string", "null"]
          required:
            - raceId
            - factors
            - createdAt

    RaceResultSummary:
      type: object
      properties:
        winnerHorseId:
          type: integer
        winnerHorseName:
          type: string
        runnerUpHorseId:
          type: ["integer", "null"]
        runnerUpHorseName:
          type: ["string", "null"]
        thirdHorseId:
          type: ["integer", "null"]
        thirdHorseName:
          type: ["string", "null"]
        recordedAt:
          type: string
        notes:
          type: ["string", "null"]
        topPickCorrect:
          type: ["boolean", "null"]
      required:
        - winnerHorseId
        - winnerHorseName
        - recordedAt

    Race:
      type: object
      properties:
        id:
          type: integer
        raceNumber:
          type: integer
        name:
          type: string
        venue:
          type: string
        distance:
          type: integer
        raceTime:
          type: string
        meetingDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, analyzing, completed, cancelled]
        surface:
          type: string
        grade:
          type: ["string", "null"]
        prize:
          type: ["string", "null"]
        horseCount:
          type: integer
        nextUpdateAt:
          type: ["string", "null"]
        lastAnalyzedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        syncedFrom:
          type: ["string", "null"]
        isToday:
          type: boolean
        isThisWeek:
          type: boolean
        dayLabel:
          type: string
        minutesToRace:
          type: ["integer", "null"]
        forecastBand:
          type: string
        prominence:
          type: number
        topPrediction:
          anyOf:
            - $ref: "#/components/schemas/PredictionSummary"
            - type: "null"
        topPredictions:
          type: array
          items:
            $ref: "#/components/schemas/PredictionSummary"
        result:
          anyOf:
            - $ref: "#/components/schemas/RaceResultSummary"
            - type: "null"
      required:
        - id
        - raceNumber
        - name
        - venue
        - distance
        - raceTime
        - status
        - surface
        - horseCount
        - createdAt
        - isToday
        - isThisWeek
        - dayLabel
        - forecastBand
        - prominence
        - topPredictions

    RaceDetail:
      allOf:
        - $ref: "#/components/schemas/Race"
        - type: object
          properties:
            horses:
              type: array
              items:
                $ref: "#/components/schemas/Horse"
            predictions:
              type: array
              items:
                $ref: "#/components/schemas/Prediction"
          required:
            - horses
            - predictions

    Horse:
      type: object
      properties:
        id:
          type: integer
        raceId:
          type: integer
        name:
          type: string
        number:
          type: integer
        jockey:
          type: string
        trainer:
          type: string
        form:
          type: string
        weight:
          type: ["number", "null"]
        currentOdds:
          type: number
        openingOdds:
          type: ["number", "null"]
        oddsMovement:
          type: string
          enum: [shortening, drifting, stable, unknown]
        scratched:
          type: boolean
        scratchReason:
          type: ["string", "null"]
        courseRecord:
          type: boolean
        distanceRecord:
          type: boolean
        trainerJockeyRecord:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - raceId
        - name
        - number
        - jockey
        - trainer
        - form
        - currentOdds
        - oddsMovement
        - scratched
        - courseRecord
        - distanceRecord
        - trainerJockeyRecord
        - createdAt

    AnalysisResult:
      type: object
      properties:
        raceId:
          type: integer
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/Prediction"
        analyzedAt:
          type: string
        nextUpdateAt:
          type: string
      required:
        - raceId
        - predictions
        - analyzedAt
        - nextUpdateAt

    PredictionWeights:
      type: object
      properties:
        id:
          type: integer
        courseForm:
          type: number
        formDistance:
          type: number
        jockeyTrainer:
          type: number
        oddsMovement:
          type: number
        history:
          type: number
        updatedAt:
          type: string
      required:
        - id
        - courseForm
        - formDistance
        - jockeyTrainer
        - oddsMovement
        - history
        - updatedAt

    UpdateWeightsBody:
      type: object
      properties:
        courseForm:
          type: number
        formDistance:
          type: number
        jockeyTrainer:
          type: number
        oddsMovement:
          type: number
        history:
          type: number
      required:
        - courseForm
        - formDistance
        - jockeyTrainer
        - oddsMovement
        - history

    CreateRaceBody:
      type: object
      properties:
        raceNumber:
          type: integer
        name:
          type: string
        venue:
          type: string
        distance:
          type: integer
        raceTime:
          type: string
        meetingDate:
          type: ["string", "null"]
        surface:
          type: string
        grade:
          type: ["string", "null"]
        prize:
          type: ["string", "null"]
      required:
        - raceNumber
        - name
        - venue
        - distance
        - raceTime
        - surface

    CreateHorseBody:
      type: object
      properties:
        name:
          type: string
        number:
          type: integer
        jockey:
          type: string
        trainer:
          type: string
        form:
          type: string
        weight:
          type: ["number", "null"]
        currentOdds:
          type: number
        openingOdds:
          type: ["number", "null"]
        courseRecord:
          type: boolean
        distanceRecord:
          type: boolean
        trainerJockeyRecord:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - name
        - number
        - jockey
        - trainer
        - form
        - currentOdds

    RecordRaceResultBody:
      type: object
      properties:
        winnerHorseId:
          type: integer
        runnerUpHorseId:
          type: ["integer", "null"]
        thirdHorseId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - winnerHorseId

    ChatMessageBody:
      type: object
      properties:
        message:
          type: string
        raceId:
          type: ["integer", "null"]
      required:
        - message

    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        raceId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - role
        - content
        - createdAt

    ChatResponse:
      type: object
      properties:
        message:
          type: string
        updatedWeights:
          anyOf:
            - $ref: "#/components/schemas/PredictionWeights"
            - type: "null"
        triggeredAnalysis:
          type: boolean
      required:
        - message
        - triggeredAnalysis

    PerformanceResultSummary:
      type: object
      properties:
        raceId:
          type: integer
        raceName:
          type: string
        meetingDate:
          type: ["string", "null"]
        topPickHorseName:
          type: ["string", "null"]
        winnerHorseName:
          type: string
        topPickCorrect:
          type: boolean
      required:
        - raceId
        - raceName
        - winnerHorseName
        - topPickCorrect

    DashboardPerformanceSummary:
      type: object
      properties:
        sampleSize:
          type: integer
        topPickWinRate:
          type: number
        placedRate:
          type: number
        averageConfidence:
          type: number
        confidenceBias:
          type: number
        factorAdjustments:
          type: object
          properties:
            courseForm:
              type: number
            formDistance:
              type: number
            jockeyTrainer:
              type: number
            oddsMovement:
              type: number
            history:
              type: number
          required:
            - courseForm
            - formDistance
            - jockeyTrainer
            - oddsMovement
            - history
        strongestEdge:
          type: ["string", "null"]
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/PerformanceResultSummary"
      required:
        - sampleSize
        - topPickWinRate
        - placedRate
        - averageConfidence
        - confidenceBias
        - factorAdjustments
        - recentResults

    WeeklyOverviewDay:
      type: object
      properties:
        date:
          type: string
        label:
          type: string
        raceCount:
          type: integer
        analyzedCount:
          type: integer
        completedCount:
          type: integer
        venues:
          type: array
          items:
            type: string
        spotlightRaceId:
          type: ["integer", "null"]
        spotlightRaceName:
          type: ["string", "null"]
        spotlightHorseName:
          type: ["string", "null"]
        spotlightConfidence:
          type: ["number", "null"]
      required:
        - date
        - label
        - raceCount
        - analyzedCount
        - completedCount
        - venues

    DashboardSummary:
      type: object
      properties:
        totalRaces:
          type: integer
        analyzedRaces:
          type: integer
        upcomingRaces:
          type: integer
        completedRaces:
          type: integer
        totalHorses:
          type: integer
        todayRaceCount:
          type: integer
        weekRaceCount:
          type: integer
        nextRaceTime:
          type: ["string", "null"]
        nextRaceVenue:
          type: ["string", "null"]
        topPick:
          type: ["string", "null"]
        topPickRace:
          type: ["string", "null"]
        venues:
          type: array
          items:
            type: string
        todayCards:
          type: array
          items:
            $ref: "#/components/schemas/Race"
        weeklyOverview:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyOverviewDay"
        performance:
          $ref: "#/components/schemas/DashboardPerformanceSummary"
      required:
        - totalRaces
        - analyzedRaces
        - upcomingRaces
        - completedRaces
        - totalHorses
        - todayRaceCount
        - weekRaceCount
        - venues
        - todayCards
        - weeklyOverview
        - performance